pick_optimization/tour_formation/requirements.txt
 # Core Python package management
pip>=23.0.0
setuptools>=68.0.0
wheel>=0.40.0

# Jupyter Notebook Support
jupyter>=1.0.0
ipykernel>=6.0.0

# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0

# Database connections
vertica_python>=1.0.0

# Optimization
gurobipy==11.0.0

# Logging and configuration
pyyaml>=6.0.0

# Cryptography (for encryption.py)
cryptography>=41.0.0

# Tabulate (for pretty-printing tables in logs)
tabulate>=0.9.0

# Code formatting and linting (recommended for code quality)
black>=23.0.0
flake8>=6.0.0

# Dependency management (useful for managing project dependencies)
pip-tools>=7.0.0

# Scientific computing and visualization
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine learning and clustering
scikit-learn>=1.3.0